requirements.txt
numpy
pandas
scikit-learn
PyMuPDF
PyPDF2
pdfplumber
pdfminer.six
spacy
